--- src/main.cpp
@@ -1,3 +1,0 @@
-auto main() -> int {
-    return 0;
-}
